split all current dfs/bfs methods into their own modules
