+++ .gitignore
@@ -3,3 +3,4 @@ Cargo.lock
 /tests/test_zome/tests/node_modules/*
 /tests/test_zome/target
 /tests/test_zome/tests/out.log
+*.dna
