+++ tests/test_zome/workdir/dna.yaml
@@ -4,7 +4,7 @@ name: "time-index-test"
 uid: 00000000-0000-0000-0000-000000000000
 properties: {
   "enforce_spam_limit": 20,
-  "max_chunk_interval": 100000,
+  "max_chunk_interval": 10000,
   "active_agent_duration_s": 300,
   "enable_signals": true,
   "enable_time_index": true,
